chore(e2e): remove artillery dependency
