+++ packages/e2e/src/index.ts
@@ -1,6 +1,7 @@
-export * from './factories';
-export * from './util';
-export * from './environment';
+export * from './address';
 export * from './defaults';
-export * from './measurement-util';
+export * from './environment';
+export * from './factories';
 export * from './load-test-scheduler';
+export * from './measurement-util';
+export * from './util';
