--- packages/e2e/test/artillery/StakePoolSearch.yml
@@ -1,23 +1,0 @@
-config:
-  # target is a required field for artillery, but it is actually
-  # ignored since we are not using artillery to perform the HTTP requests.
-  target: 'http://localhost'
-  http:
-    timeout: 180
-  phases:
-    - name: 'Stake Pool Search'
-      duration: 1
-      arrivalCount: 1
-    - name: 'Stake Pool Search Queries'
-      duration: '{{$processEnvironment.TEST_DURATION_IN_SECS}}'
-      arrivalCount: '{{$processEnvironment.VIRTUAL_USERS_COUNT}}'
-  processor: './StakePoolSearch.ts'
-
-scenarios:
-  - name: 'Browse'
-    flow:
-      - function: 'healthCheck'
-      # Loop over result pages
-      - loop:
-          - function: 'performQuery'
-        whileTrue: 'moreResults'
